Fix: signUpItemSlice => createItemsSlice로 변경 및 기능 추가
